microsoft-fabric/integracio_per_consultar_dades/exemples/graphql/requirements.txt
azure-identity
pip-system-certs
msal 
requests
pandas
